Produkti tiek pievienoti grozam, ar x pogu tos var izdzest un paradas pazinojums ka tie ir izdzesti.

+++ database/migrations/2022_11_25_064603_create_create_posts_table.php
@@ -16,7 +16,7 @@ class CreateCreatePostsTable extends Migration
         Schema::create('create_posts', function (Blueprint $table) {
             $table->id();
             $table->string('product_name', 64);
-            $table->float('product_price');
+            $table->integer('product_price');
             $table->string('product_description');
             $table->unsignedBigInteger('author_id');
             $table->timestamps();
